Carbon int to double #155

See merge request cse1105/2018-2019/oopp-group-4/template!118

+++ src/main/java/server/Action.java
@@ -6,18 +6,23 @@ package server;
 public class Action {
     private String token;
     private String action;
-    private int value;
-    private int carbonReduced;
-    private int carbonProduced;
-    private long date;
-
+    private int    value;
+    private double carbonReduced;
+    private double carbonProduced;
+    private long   date;
+    
     /**
      * empty constructor.
      */
     public Action() {
     }
-
-    public Action(int carbonProduced, int carbonReduced) {
+    
+    /**
+     * constructor containing only carbon values
+     * @param carbonProduced kilograms of carbon produced
+     * @param carbonReduced  kilograms of carbon produced
+     */
+    public Action(double carbonProduced, double carbonReduced) {
         this.carbonProduced = carbonProduced;
         this.carbonReduced = carbonReduced;
     }
@@ -44,8 +49,7 @@ public class Action {
      * @param carbonProduced carbon produced
      * @param carbonReduced  carbon saced
      */
-    public Action(String token, String action, int value, int carbonReduced, int carbonProduced) {
-
+    public Action(String token, String action, int value, double carbonReduced, double carbonProduced) {
         this.token = token;
         this.action = action;
         this.value = value;
@@ -63,8 +67,7 @@ public class Action {
      * @param carbonReduced  carbon saced
      * @param date           date added (in miliseconds)
      */
-    public Action(String token, String action, int value,
-                  int carbonReduced, int carbonProduced, long date) {
+    public Action(String token, String action, int value, double carbonReduced, double carbonProduced, long date) {
         this.token = token;
         this.action = action;
         this.value = value;
@@ -82,7 +85,7 @@ public class Action {
      * @param carbonReduced  carbon saced
      * @param date           date added (in miliseconds)
      */
-    public Action(String action, int value, int carbonReduced, int carbonProduced, long date) {
+    public Action(String action, int value, double carbonReduced, double carbonProduced, long date) {
         this.action = action;
         this.value = value;
         this.carbonReduced = carbonReduced;
@@ -143,14 +146,14 @@ public class Action {
      *
      * @return carbonReduced
      */
-    public int getCarbonReduced() {
+    public double getCarbonReduced() {
         return carbonReduced;
     }
 
     /**
      * setter for carbonReduced.
      */
-    public void setCarbonReduced(int carbonReduced) {
+    public void setCarbonReduced(double carbonReduced) {
         this.carbonReduced = carbonReduced;
     }
 
@@ -159,14 +162,14 @@ public class Action {
      *
      * @return carbonProduced
      */
-    public int getCarbonProduced() {
+    public double getCarbonProduced() {
         return carbonProduced;
     }
 
     /**
      * setter for Value.
      */
-    public void setCarbonProduced(int carbonProduced) {
+    public void setCarbonProduced(double carbonProduced) {
         this.carbonProduced = carbonProduced;
     }
